Fix VS2019 build error: cast CHAR16 to UINT8 in GetControlCharDescription
- Fix C2220/C4244 warning treated as error on line 271
- Cast KeyData->Key.UnicodeChar to UINT8 when calling GetControlCharDescription
- Also cast in DEBUG statements for consistency
- Cast is safe since values are already checked to be in range 0x01-0x1F

+++ Platform/SimpleTextInputExCompatDxe/SimpleTextInputExCompat.c
@@ -268,15 +268,15 @@ CompatReadKeyStrokeEx (
       KeyData->KeyState.KeyShiftState = EFI_LEFT_CONTROL_PRESSED;
 
       // Log control character using lookup table
-      CHAR8  *Description = GetControlCharDescription (KeyData->Key.UnicodeChar);
+      CHAR8  *Description = GetControlCharDescription ((UINT8)KeyData->Key.UnicodeChar);
       if (Description != NULL) {
         DEBUG ((DEBUG_INFO, "STX: %a detected\n", Description));
       } else {
         DEBUG ((
           DEBUG_INFO,
           "STX: Control character 0x%02X (Ctrl+%c) detected\n",
-          KeyData->Key.UnicodeChar,
-          KeyData->Key.UnicodeChar + 0x40
+          (UINT8)KeyData->Key.UnicodeChar,
+          (UINT8)KeyData->Key.UnicodeChar + 0x40
           ));
       }
     } else {
